backend/requirements.txt
fastapi==0.99.0
uvicorn[standard]==0.22.0
sqlalchemy==1.4.49
databases==0.6.0
pydantic==1.10.9
psycopg2-binary==2.9.7
alembic==1.11.1
python-dotenv==1.0.0
gunicorn==21.2.0
